refactor: extract project member-related service calls into a dedicated ` ProjectMemberService `
